Adding files which were made to finish the CRUD work. TODO - add the CRUD to these aspects, finish refactoring, upload then merge to main.
