+++ shelves.go
@@ -1,0 +1,9 @@
+package main
+
+import "gorm.io/gorm"
+
+type Shelf struct {
+	gorm.Model
+	UserID uint
+	Name   string
+}
